requirements.txt
aiogram>=3.0.0
gspread>=5.10.0
oauth2client>=4.1.3
pandas>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
